refactor: moved tests away from unit tests and into integration tests
