+++ src/matrix/elementwise_op/addop.rs
@@ -334,106 +334,6 @@ where
 
 mod tests
 {
-
-    use super::*;
-
-    #[test]
-    fn test_add_smatrix()
-    {
-        let matrix1 = SMatrix::<i32, 2, 2>::from_value(1);
-
-        let matrix2 = SMatrix::<i32, 2, 2>::from_value(10);
-
-        let matrix3 = SMatrix::<i32, 2, 2>::from_value(100);
-
-        let matrix4 = SMatrix::<i32, 2, 2>::from_value(1000);
-
-        let matrix5 = SMatrix::<i32, 2, 2>::from_value(10000);
-
-        let matrix6 = SMatrix::<i32, 2, 2>::from_value(100000);
-
-        let matrix7 = SMatrix::<i32, 2, 2>::from_value(1000000);
-
-        let mut matrix8 = SMatrix::<i32, 2, 2>::default();
-
-        matrix8 = (&matrix7 + (&matrix4 + &matrix5) + (&matrix1 + &matrix2 + &matrix3) + &matrix6)
-            .evals();
-
-        let exp_value: i32 = 1000000 + (1000 + 10000) + (1 + 10 + 100) + 100000;
-
-        for val in &matrix8
-        {
-            assert_eq!(*val, exp_value);
-        }
-    }
-
-    #[test]
-    fn test_add_dmatrix()
-    {
-        let matrix1 = DMatrix::<i32>::from_value(2, 2, 1);
-
-        let matrix2 = DMatrix::<i32>::from_value(2, 2, 10);
-
-        let matrix3 = DMatrix::<i32>::from_value(2, 2, 100);
-
-        let matrix4 = DMatrix::<i32>::from_value(2, 2, 1000);
-
-        let matrix5 = DMatrix::<i32>::from_value(2, 2, 10000);
-
-        let matrix6 = DMatrix::<i32>::from_value(2, 2, 100000);
-
-        let matrix7 = DMatrix::<i32>::from_value(2, 2, 1000000);
-
-        let mut matrix8 = DMatrix::<i32>::zeros(2, 2);
-
-        matrix8 = (&matrix7 + (&matrix4 + &matrix5) + (&matrix1 + &matrix2 + &matrix3) + &matrix6)
-            .evald();
-
-        let exp_value: i32 = 1000000 + (1000 + 10000) + (1 + 10 + 100) + 100000;
-
-        for val in &matrix8
-        {
-            assert_eq!(*val, exp_value);
-        }
-    }
-
-    #[test]
-    fn test_add_scalar_smatrix()
-    {
-        let matrix1 = SMatrix::<i32, 2, 2>::from_value(10);
-
-        let matrix2 = SMatrix::<i32, 2, 2>::from_value(100);
-
-        let mut matrix4 = SMatrix::<i32, 2, 2>::default();
-
-        matrix4 = (4i32 + (2i32 + &matrix1) + (&matrix2 + 3i32) + 5i32).evals();
-
-        let exp_val = 4 + (2 + 10) + (100 + 3) + 5;
-
-        for val in &matrix4
-        {
-            assert_eq!(*val, exp_val);
-        }
-    }
-
-    #[test]
-    fn test_add_scalar_dmatrix()
-    {
-        let matrix1 = DMatrix::<i32>::from_value(2,2, 10);
-
-        let matrix2 = DMatrix::<i32>::from_value(2,2, 100);
-
-        let mut matrix4 = DMatrix::<i32>::zeros(2,2);   
-
-        matrix4 = (4i32 + (2i32 + &matrix1) + (&matrix2 + 3i32) + 5i32).evald();
-
-        let exp_val = 4 + (2 + 10) + (100 + 3) + 5;
-
-        for val in &matrix4
-        {
-            assert_eq!(*val, exp_val);
-        }
-    }
 }
 
 //}}}
