+++ src/matrix/elementwise_op/subop.rs
@@ -1,7 +1,6 @@
-//! Short Description of module
+//! This mdule implements the element-wise subtraction operation for the `SMatrix` and `DMatrix` 
+//! types.
 //!
-//! Longer description of module
-
 //--------------------------------------------------------------------------------------------------
 
 //{{{ crate imports
@@ -10,12 +9,10 @@ use super::super::smatrix::*;
 use super::common::{BinOp, BinopExpr, SubOp};
 use crate::apply_for_all_types;
 use crate::common::*;
-
 //}}}
 //{{{ std imports
 use std::fmt;
 use std::ops::Sub;
-
 //}}}
 //{{{ dep imports
 //}}}
@@ -320,92 +317,5 @@ where
 mod tests
 {
 
-    use super::*;
-
-    #[test]
-    fn test_sub_smatrix()
-    {
-        let matrix1 = SMatrix::<i32, 2, 2>::from_value(1);
-
-        let matrix2 = SMatrix::<i32, 2, 2>::from_value(10);
-
-        let matrix3 = SMatrix::<i32, 2, 2>::from_value(100);
-
-        let matrix4 = SMatrix::<i32, 2, 2>::from_value(1000);
-
-        let matrix5 = SMatrix::<i32, 2, 2>::from_value(10000);
-
-        let matrix6 = SMatrix::<i32, 2, 2>::from_value(100000);
-
-        let mut matrix7 = SMatrix::<i32, 2, 2>::default();
-
-        matrix7 = ((&matrix4 - &matrix5) - (&matrix1 - &matrix2 - &matrix3) - &matrix6).evals();
-
-        let exp_value: i32 = (1000 - 10000) - (1 - 10 - 100) - 100000;
-
-        for val in &matrix7
-        {
-            assert_eq!(*val, exp_value);
-        }
-    }
-
-    #[test]
-    fn test_sub_dmatrix()
-    {
-        let matrix1 = DMatrix::<f64>::from_value(2, 2, 1.0);
-        let matrix2 = DMatrix::<f64>::from_value(2, 2, 10.0);
-        let matrix3 = DMatrix::<f64>::from_value(2, 2, 100.0);
-        let matrix4 = DMatrix::<f64>::from_value(2, 2, 1000.0);
-        let matrix5 = DMatrix::<f64>::from_value(2, 2, 10000.0);
-        let matrix6 = DMatrix::<f64>::from_value(2, 2, 100000.0);
-        let mut matrix7 = DMatrix::<f64>::zeros(2, 2);
-
-        matrix7 = ((&matrix4 - &matrix5) - (&matrix1 - &matrix2 - &matrix3) - &matrix6).evald();
-
-        let exp_value: f64 = (1000.0 - 10000.0) - (1.0 - 10.0 - 100.0) - 100000.0;
-
-        for val in &matrix7
-        {
-            assert_eq!(*val, exp_value);
-        }
-    }
-
-    #[test]
-    fn test_sub_scalar()
-    {
-        let matrix1 = SMatrix::<i32, 2, 2>::from_value(10);
-
-        let matrix2 = SMatrix::<i32, 2, 2>::from_value(100);
-
-        let mut matrix4 = SMatrix::<i32, 2, 2>::default();
-
-        matrix4 = (4i32 - (2i32 - &matrix1) - (&matrix2 - 3i32) - 5i32).evals();
-
-        let exp_val = 4 - (2 - 10) - (100 - 3) - 5;
-
-        for val in &matrix4
-        {
-            assert_eq!(*val, exp_val);
-        }
-    }
-
-
-    #[test]
-    fn test_sub_scalar_dmatrix()
-    {
-        let matrix1 = DMatrix::<f64>::from_value(2, 2, 10.0);
-        let matrix2 = DMatrix::<f64>::from_value(2, 2, 100.0);
-        let mut matrix4 = DMatrix::<f64>::zeros(2, 2);
-
-        matrix4 = (4.0 - (2.0 - &matrix1) - (&matrix2 - 3.0) - 5.0).evald();
-
-        let exp_val = 4.0 - (2.0 - 10.0) - (100.0 - 3.0) - 5.0;
-
-        for val in &matrix4
-        {
-            assert_eq!(*val, exp_val);
-        }
-    }
 }
-
 //}}}
